Lower the mutation rate
In the current way of mutating, when using a 0.2 mutation rate it will lead to 98.8% chance of making a mutation. This is a very high value! we want to keep most crossover chromosomes to NOT be mutated.
Using the mutation rate as 0.005 will make that chance to be around 1% .
However, this doesn't have much of an impact after the last change (extending last generation).